All of Dc.
Hm. I'll be keeping to shorter commits from now on. It makes it easier
to keep a log of the changes I've made to the various sources.

Removing some of the Dc subtitles that just contained nothing.
Adding all of the ones that do.

Significant changes to the timing for quite a few of Dc, as well as
adding a few vocalizations that were missing.
Unfortunately I didn't write down exactly which.